# Handle computers with multiple processors