We were collecting the energy consumption of a single processor package.

This patch collects the energy consumption of each package and sum it
up to get the total energy consumption.

The code was based on the PAPI RAPL code [1][2].

[1] http://icl.cs.utk.edu/papi/
[2] http://gist.github.com/luisgabriel/f91656a6662e21a2ff9c