@luisgabriel 2015-06-17
Just tested on linuxapps02:

```
====> BEFORE PATCH
benchmarking fasta #7
measurement took 64.88 s
analysing with 1000 resamples
measurement overhead 36.77 μs
bootstrapping with 3 of 4 samples (75%)
time                 4.073 s    (3.323 s .. 5.143 s)
                     0.990 R²   (0.987 R² .. 1.000 R²)
mean                 4.478 s    (4.181 s .. 4.679 s)
std dev              301.9 ms   (0.0 s .. 348.4 ms)
energy:              0.947 R²   (0.927 R² .. 1.000 R²)
  iters              69.799     (40.278 .. 112.602)
  y                  38.665     (NaN .. 140.326)
variance introduced by outliers: 20% (moderately inflated)


====> AFTER PATCH
Total packages found: 2
benchmarking fasta #7
measurement took 67.16 s
analysing with 1000 resamples
measurement overhead 56.31 μs
bootstrapping with 3 of 4 samples (75%)
time                 4.400 s    (4.292 s .. NaN s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 4.502 s    (4.480 s .. 4.515 s)
std dev              19.88 ms   (0.0 s .. 22.77 ms)
energy:              1.000 R²   (1.000 R² .. 1.000 R²)
  iters              233.652    (227.020 .. 237.834)
  y                  4.492      (-7.887 .. 13.743)
variance introduced by outliers: 19% (moderately inflated)
```